+++ Modules/ADUtils/UICollectionViewCell+PreferredLayoutHeight.swift
@@ -15,7 +15,26 @@ public extension UICollectionViewCell {
 
      - parameter fittingWidth: The biggest width the view can get
      */
+    @available(*, deprecated, message: "Use `ad_preferredCellLayoutSize(fittingWidth:)` directly instead")
     func ad_preferredCellLayoutHeight(fittingWidth: CGFloat) -> CGFloat {
-        return contentView.ad_preferredLayoutHeight(fittingWidth: fittingWidth)
+        return ad_preferredCellLayoutSize(fittingWidth: fittingWidth).height
+    }
+
+    /**
+     Provides the preferred layout size for the collection view cell, this is the smallest size the view and its content can fit. You should populate the view before calling this method.
+
+     - parameter fittingWidth: The biggest width the view can get
+     */
+    func ad_preferredCellLayoutSize(fittingWidth: CGFloat) -> CGSize {
+        return contentView.ad_preferredLayoutSize(fittingWidth: fittingWidth)
+    }
+
+    /**
+     Provides the preferred layout size for the collection view cell, this is the smallest size the view and its content can fit. You should populate the view before calling this method.
+
+     - parameter fittingHeight: The biggest height the view can get
+     */
+    func ad_preferredCellLayoutSize(fittingHeight: CGFloat) -> CGSize {
+        return contentView.ad_preferredLayoutSize(fittingHeight: fittingHeight)
     }
 }