+++ Modules/ADUtils/UITableViewCell+PreferredLayoutHeight.swift
@@ -16,6 +16,6 @@ public extension UITableViewCell {
      - parameter fittingWidth: The biggest width the view can get
      */
     func ad_preferredCellLayoutHeight(fittingWidth: CGFloat) -> CGFloat {
-        return contentView.ad_preferredLayoutHeight(fittingWidth: fittingWidth)
+        return contentView.ad_preferredLayoutSize(fittingWidth: fittingWidth).height
     }
 }